Develop

+++ app/controllers/cart_items/orders_controller.rb
@@ -12,19 +12,31 @@ class CartItems::OrdersController < ApplicationController
     @order.user_id = current_user.id
     @order.shop_id = @shop.id
     @order.numbering_reserve_number(@order.shop.target_date_order_count(@order.takeaway_datetime))
-    if @order.save
-      @cart_items = current_user.cart_items
-      @cart_items.each do |cart_item|
-        OrderItem.create(
-          order_id: @order.id,
-          menu_id: cart_item.menu.id,
-          menu_name: cart_item.menu.name,
-          menu_price: cart_item.menu.price,
-          menu_amount: cart_item.amount
-        )
-        cart_item.destroy
+    if @order.is_business_time_order?
+      if @order.save
+        @cart_items = current_user.cart_items
+        @cart_items.each do |cart_item|
+          OrderItem.create(
+            order_id: @order.id,
+            menu_id: cart_item.menu.id,
+            menu_name: cart_item.menu.name,
+            menu_price: cart_item.menu.price,
+            menu_amount: cart_item.amount
+          )
+          cart_item.destroy
+        end
+        OrderMailer.send_user_order(@order).deliver_later
+        redirect_to orders_fix_path
+      else
+        flash[:danger] = "注文できませんでした。入力内容を確認してください"
+        render "new"
       end
-      redirect_to orders_fix_path
+    else
+      flash.now[:danger] = "受け取り希望時間をご確認ください"
+      render :new
+    end
+    if @order.save
+      
     else
       flas.now[:danger] = "注文できませんでした。受け取り時間を確認してください。"
       render :new
